对服务器端做适配
1、降低Java版本为8，这是由于服务器的Java版本就是8，然后我懒得给服务器升级Java版本
2、增加gitignore的一行，把out文件夹给剔除在git之中
3、发现有些数据使用int无法完整表示，我们提升为long
4、服务器的8080端口被占用，我们改为18080端口

+++ pm-project-backdoor/src/main/java/cn/edu/xjtu/stu/orangesoft/backdoor/pojo/Student.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Component;
 public class Student {
     private int StudentUserID;
     private String StudentClass;
-    private int StudentID;
+    private long StudentID;
     private int TeamID;
 }
